Move group router forms into view definition.
